feat(phase-5.3b): tile clipping grid rewrite brings Stage3 PSNR 72→102 dB; add contract-validated compare_psnr.py
- dng_warp_halide.cpp: rebuild buildTileClippingGrid with per-plane corner
  scan; fix frac_idx=0 on clip; Stage3 PSNR 72.44→102.71 dB (+30 dB),
  Stage3 time 224→136 ms (−39%), total decode 718→415 ms (2× vs baseline)
- DngWarpGenerator.cpp: sync Halide AOT kernel clip/frac behaviour
- compare_psnr.py: full rewrite — contract checks (width/height/planes/
  pixel_size/layout/buffer_size) before PSNR; per-channel MAE/maxAbs/
  location; uint16 headless raw support; abort exit(2) on any FAIL
- run_decode_matrix.py / compare_psnr.py: add YAML file-index headers
  (rule.md Appendix A) so AI agents can locate functions in <100 lines
- docs: sync memory/task/handover/plan/unit_test/file_index with latest
  PSNR numbers and performance data; add task log Section 7

+++ dng_processor/native/tests/test_decode.cpp
@@ -244,7 +244,7 @@ bool applySingleWarpRectilinearInterleaved(dng_negative& negative,
 
     WarpRectilinearParams params;
     if (!extractWarpRectilinearParams(warpOpcode,
-                                      static_cast<float>(negative.PixelAspectRatio()),
+                                      negative.PixelAspectRatio(),
                                       params)) {
         return false;
     }
